Record my feeble attempt of extracting the mergeinfo from the SVN repo
svn_fs_get_mergeinfo() seems to return all mergeinfo, so instead of
trying to understand the API better, shell out to svn(1), yolo style,
and at least record the diff in properties in the log. This data
should be useful in potentially manually curating the list of merges
from vendor branches to head that we want to record as proper git
merges.

Sadly, the output also contains all other property changes, like
svn:keywords or mimetypes. It blows up the logging output to almost 1GB.

It's a first step towards fixing #1 properly.