Hoist the printing of potential mergeinfo up a layer or two.
This way, we can only print it once per SVN revision (not per path
exported) and also will only print it should there have not already been
a merge recorded.

At least that's the idea, though it fails on r179468, which has
mergeinfo and *does* shift a whole bunch of stuff around, but at least
it does not merge anything from one tree to another, and it lists the
mergeinfo like so, which I think means it has "empty" mergeinfo added.

  Property changes on: vendor/gcc/2.95.3-test1/gcc
  ___________________________________________________________________
  Added: svn:mergeinfo
  ## -0,0 +0,0 ##