Checkpoint work on making svn:mergeinfo merges work.
This detects about 47889 commits that were previously ignored. Of
course, most of these are merges into stable/X, that we don't "merge"
but instead record as cherry-picks only. Disregarding merges into releng
or stable, we're left with 1705 of such unhandled merges. Some of them
are within a branch (flattening vendor trees, for example), but all of
them will need to be inspected and handled.

The current heuristic manages to handle only 22 of them.

This is a major step towards fixing #1.