Use the actual rule parsing to figure out where the merge went (and also from where it came).
Reorder things heavily to whittle down the number of revisions to act
on.

This now produces about 1949 merge commits for the 3024 non-empty
mergeinfo revisions.

It cuts the "obvious" missing MFV merges on master down from 1600 to
about 800, though many of them are in the CVS days anyway. For the
SVN-only history the missing merge commits went down from 1100 to 400.

Pretty good step towards fixing #1.