Merge the sendmail dist and dist-old vendor branches
This adds a new keyword to strip a certain prefix of the resulting path.
This is required as a more deep "match" regexp would then result in the
branchNotes not ending in vendor/sendmail/dist-old and then a later
merge-from that branch wouldn't find it. So we need to end the match
early.

While this basically works out, the problem is that the later import
gets _added_ to the existing data, it doesn't replace it. Meaning we'll
have quite the leftover schmutz. The flattening of the vendor tree then
completely deletes the old path, clearing up the mismatch.

Note sure we want to run with this. Ideally we could tell svn2git to
delete an arbitrary list of paths in a branch in a certain revision.
It would also help with the ipfilter repo-copy among others.

Someone should code this up ...