Add the source code, as shipped per release, into the history
This fetches the old, checked out sources from the archive server
and stuffs them into the RELEASE tag history as an additional commit and
tag.

Not all sources are available and we're not publishing 1.0-RELEASE
either. Of note is that releases prior to 5.3 do not have crypto/ and
secure/ subdirs on the archive server, so this is missing.

The nice thing about this is that it brings back the expanded CVS Id
tags (and SVN revision tags) and closes issue #8. It can be considered
a minimal solution for issue #7 as well.