Compare almost every frigging vendor branch and tag between SVN and git
The shell script will soon collapse under the weight of hacks, but so be
it. It identified half a dozen of issues with the conversion (fixed in
the rules) and another half dozen that still needs followups but
otherwise they are somewhat minor problems.

This mostly fixes issue #6, but some more tags and branches need fixups.