Fetch everything from the archive, including 1.x releases.
We should be ok publishing this, according to some internal emails.

The 2.x ISOs don't work with bsdtar, so use 7z to extract the src
tarballs.

The 1.x commit messages and dates are WIP and I figured to stash them
all under stable/1, even though we don't cut releases from stable but
releng instead, but having releng/1.0 and releng/1.1 seemed stupid for
the 3 commits.

This mostly fixes issue #8, except for all the pesky details of naming
and commit dates and messages, etc. pp.