Solve the mystery of the disconnected tags
The `prefix` action was applied to the source of the branchpoint, but
not to the resulting tag. With prefixes not matching, we can't do a
plain fork. Turn the warning about this into something fatal. It
otherwise doesn't happen for our repo anyway.

Also, fix the case where we want to have vendor/sendmail/dist end up as
vendor/sendmail directly. operator[] is evil and was creating a default
object when the mergeinfo parser was returning the fact that something
was merged *from* vendor/sendmail/dist.

This seems very brittle and we should probably avoid doing this and
instead rename the refs post-conversion. Or tell all vendor software
maintainers to push a branch called vendor/foo and delete the
vendor/foo/dist one.

This commit is mainly to show what could've been possible.