Implement renames, use them to fix up the ipfilter repo-moves (at least in the vendor area)
It turns out we can't just use the `delete` action, as it will always
happen first (has to, otherwise the tree comes out totally different).
So provide a rename action to do the proper renames once all file
modifications have been written to the stream. Also handle renames with
an empty target or /dev/null as deletions that happen after all
modifications have been written.

Use this all together to fix up the historical mess that is the ipfilter
vendor area. There are still 2 files that should probably be removed as
they have likely been repo-moved around, but first things first.

This is part of issue #5