Handle more mergeinfos where we have a whole set of more or less identical merge information.
This for example fixes the merges from
vendor/illumos+vendor-sys/illumos, as they end up in the same git branch
anyway. It also fixes the mergeinfo for the llvm imports, where
mergeinfo is added to all subdirs like clang, llvm, lld, compiler-rt,
etc. but they are all subdirs of 1 single branch these days, so emiting
a single merge commit is easy.

Some of the older LLVM merges are not that simple and it will require
bubbling up the set of mergeinfo in a followup commit.