Remove the 16 parents limit, not sure why that was ever there?
This will actually make the 3 revisions that are affected worse, but
we'll need a proper fix for this case anyway, we can't just special
case the cvs2svn tag case, as regular svn tags will produce the same
garbage diff.

This is an analysis of issue #16 and basically a dupe of issue #9.