Twist the branch creation logic to not create stupid vendor tags.
This fixes issue #9 and is another step to fixing issue #5

It has a downside though, in that now a handful of vendor tags come out
slightly different. Some of the diffs are benign and could be considered
intended anyway, but the gcc tags are now significantly altered from the
SVN tags.