Extend the dont-merge-or-branch logic to merges as well
Use it to stop the vendor branch for pppd to record the merge from
/vendor/CSRG/dist as a proper merge. This otherwise pulls in all the
history of the initial imports, because, well the CSRG "vendor" is kinda
special here.

This also happens for jthorpe (there is just 1 commit) and for
misc-GNU/dist*, but that will be handled later. It also happens to
vendor/unknown, but most of the initial revisions also hit that one, so
not recording the branch there seems weird also.

Also fix up the history around vendor SGI, which is basically routed and
routed only. It was imported into dist, then dist2 + dist_v_2_21 at the
same time with bits and pieces going here and there. Fix this all up
nicely to have a spliced history.