Our FTP archive of ISO images is a lie.
Note down that some of these ISO images have been created long after the
actual release was done, as can be seen by the repo-copied files that
they have (but that they shouldn't have).

At least 5.5 is trustworthy, as it doesn't have the ppp-user → ppp
repo-copy yet. The 5.0 -- 5.4 images are garbage for that, but at least
you get the CVS ID keywords, so there's that.

Swap a and b, so the diff of CVS to git is better readable and rename it
to _cvs from _shipped, as it wasn't actually shipped as such.