Fix the branchpoint skip logic in the most hacky way I could quickly come up with.
While most revisions and branches were fine with the previous logic,
some of the branches are created by copying over head, and then deleting
a bunch of stuff, e.g. r66901. This doesn't work well when we toss out
the parent commit altogether.

Instead, we need to tell fast-import which tree to start out with. While
we could make use of marks here to avoid the hard-coding of tree hashes,
this would be a) more invasive and b) the tree hashes are not supposed
to change anyway (unless I fix some more repo-copies, of course).