Fix a whole lot of tags to point to some commit on the mainline
These have all been manually inspected and sometimes we just need to
skip some "empty" svn revision, more often we need to rewrite the tag,
as the CVS tag was slipped to include some file cleanups usually. So the
SVN tag predates the file deletions on the mainline branch and we then
just repoint the tag.

This is only done, if the cleanup was done very soon after the initial
tag, sometimes this is being done a year later or so, for which we don't
move the git tag, as that would be too much time travel for my liking.

This is a major step to fixing issue #9