Record the branchpoint of various releng branches closer to the actual truth.
What happened before is that the old tag slips or repo-copy cause the
branchpoint far in the past, but then only 1-2 files make it onto the
branch at first. Now, the branchpoint get's recorded with the commit
that brings back all the files into the tree.

This unbreaks e.g. `blame` for the releng branches.