Fix tags for binutils, gcc, file and v4l.
The branch-bump heuristic is too eager for them and resulted in
different tree objects. A manual branchpoint rule now overrides this. We
still need to rejigger some tags during the cleanup pass.

This should resolve the remaining diffs against the SVN tags and results
in a meaningful git branch history.

This closes issue #5.