Skip a handful of cvs2svn commits that are immediately (most of the time) overwritten by an actual commit.
This drives down the code lines at HEAD that blame would associate with
cvs2svn from 6.6k down to 182 or so.

In the case of grep, we need to very carefully slice out the useful bits
from the SVN revision that created the tag (!) and also avoid it
deleting most existing bits. This doesn't work for the AUTHORS file
though. usr.bin/ppp and the MP branch are the remaining problematic
areas.