Undo the ppp phantom chase
Turns out, according to at least CVS commit mail like this one:
https://docs.freebsd.org/cgi/getmsg.cgi?fetch=1079717+0+archive/2002/cvs-all/20020616.cvs-all
both files did exist in June 2002, so there likely was a repo-move done
that clobbered one of the files and that's why stuff comes out with a
bogus history.

Which at least means the ISO images aren't as bad as I thought and we
have that content in the tree as sort of a canonical archive (minus the
crypto stuff, of course).