gsm2 : Git to Subversion Mirror
General command to manage the mirroring from subversion to git. It
uses lock files as appropriate to ensure that we don't have two
updaters at the same time in case things run too long, etc.

Commands:
    mirror	updates the freebsd-src repo and then updates the branches
    		we are doing. It does all branches in the config file
		git2svn.conf.
    branch-init	Initializes a new branch. This sets up the repo and starts
    		keeps track of the tags needed to make things work.
    clone-init	Do the intial clone of the repo. This is usually done
    		only once. It creates a repo in 'mirror' mode so that
		things like notes that we rely on are brought forward
		(we rely on them to find the end of the branches we're
		mirroring into).

All config info is from the git2svn.conf file:
    svnurl	URL for svn repo (usually a file:/// spec)
    gitrul	Upstream repo we're replaying from
    branches	current configured branches (note: these are internal
    		names, not branch names)
    gitpath	where we place the clone of giturl
    $branch_name Actual name of the branch, same in both git and svn.
    $branch_path local path to checked-out git tree for this branch.

There are -locked versions of these commands. You can give -b on the
command-line before the command, maybe multiple times, to run subsets
of branches.

	gsm2 [ flags ] cmd

Note: the code parses -n for dryrun, but then ignores it.

I've run this for a little while now and it seems OK. There's no
provision at the current time to 'restart' the mirroring process
should something go awry in the middle. Should the need for that
arise, likely the best bet is to run branch-init, tweak
refs/remotes/upstream/${branch}-last-update by hand to the last
mirrored version in the repo, then run mirror by hand with one branch
(during all this, the cron/periodic running should be disabled).
Should this be needed often, I suspect we're better off fixing the
reason for that, but it may be possible to sift through the mirrored
branch to hook this up automatically.