Speed up conversion by re-using old pack files
As we're pushing upstream (which I think will result in a bunch of loose
objects for our incremental pushes) stop wasting time on finding the
best compression and simply re-use the pack files from the previous run.
This also means the conversion won't have write out packfiles, as little
changes.