WIP
Track which notes we've emitted for certain commit objects. Sometimes
the same commit gets tagged twice, and then we don't want to clobber the
note object with new content, but instead append the content.

This should fix doc notes for the release/2.1.5 et al. commits, which all
point to the same git commit:

  % git show -s release/2.1.5
  tag release/2.1.5
  Tagger: Doc Manager <doceng@FreeBSD.org>
  Date:   1996-07-16 16:01:36 +0000

  Create tag '2.1.5'.

  commit b6428154c7536ae3960eb29582e28f6e3eec9b61 (tag: release/2.1.7, tag: release/2.1.6, tag: release/2.1.5, releng/2.1.0)
  Author: Jordan K. Hubbard <jkh@FreeBSD.org>
  Date:   1996-07-16 16:01:35 +0000

      Might as well update the handbook while I'm at it - some sections have
      been added.

  Notes:
      svn path=/branches/RELENG_2_1_0/; revision=428
      svn path=/release/2.1.5/; revision=429; tag=release/2.1.5
      svn path=/release/2.1.6/; revision=430; tag=release/2.1.6
      svn path=/release/2.1.7/; revision=431; tag=release/2.1.7

This was done as part of fixing issue #22, but isn't absolutely required
for that.