Fix the notes ordering properly by delaying them till the end.
This allows us to insert the annotated tags at the very end, but at the
right place in the overall commit stream.

This produces a different hash for doc's refs/notes/commits, as
fast-import decides to create objects differently aa/bcdef vs. aa/bc/def
etc; in the end it settles on the same tree both ways though.

This is fine.

This should produce a stable output for doc and ports, but not yet for
src as we are still rewriting the tags after the conversion was done and
using the naive reordering script runs into merge conflicts. Two down,
one to go for issue #22.