Snapshot an almost working version
This requires to use a patched git-fast-import and produce a notes-depth
of constant 2 always in function convert_num_notes_to_fanout. This is
easy to do and means the tree of refs/notes/commits doesn't undergo the
massive rewrite from 0 to 1 to 2 levels of depth.

We also need to have the notes commit have the date of the target
commit, not the date of the tag, so that it sorts properly. At least for
vendor/groff/1.17/; revision=75586

Though I fear we actually need to sort and take the newest date, let's
see during a trial run.