Rejiggle vendor heads and tags to be user visible
Also snip of the /dist, as they no longer share a common namespace
(except of course the view where the prefix clashes).

While here, I found that I missed a vendor-sys/ ref, due to usually
grepping for vendor/ to not show all the other gunk. FML.

At least provide a replace commit to patch the ipfilter branch back
together and push that (people will need to specifically pull down the
replace, however).

For lack of a better name, name the second branch with -old. We could
also delete it fully (as the graft is in place, so it won't be GC'ed).